Externalise Worldpay code into concerns
This is an initial refactoring to the code that produces the Worldpay xml and send Worldpay requests. This only externalise common bits that can be reused with requests for refunds.
